requirements.txt
moviepy==1.0.3
protobuf==3.12.4
pydub==0.25.1
python-dotenv==1.0.0
pytube==15.0.0
google-cloud-speech
google-cloud-storage
google-cloud-texttospeech
ffmpeg-python
